Fix deadlock exception when no features are found
When calling `clone_device_descriptors_sub` the
library will check that the destination pointer is
not null and if it is exit the program. This causes
an exception to avoid deadlocking the program.

In the case of no extracted keypoints the pointer
will be null and instead of exiting and causing a
potential deadlock we just want to skip this section
entirely.